refac: criado pasta de dependencias do puzzle, renomeado o arquivo relacionado a timer e adicionado conteúdo ao diretório
