+++ lib/Puzzle/puzzle.dart
@@ -1,10 +1,65 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:puzzle_mobile/timer_controls.dart';
-import 'package:puzzle_mobile/timer.dart';
+import 'package:puzzle_mobile/Puzzle/settings.dart';
+import 'package:puzzle_mobile/Puzzle/timer.dart';
 
+class Canvas extends StatelessWidget {
+  const Canvas({super.key});
 
+  @override
+  Widget build(BuildContext context) {
+    final timer = context.watch<TimerController>();
+    final puzzle = context.watch<PuzzleController>();
+
+    return GridView.count(
+      crossAxisCount: puzzle.level,
+      padding: EdgeInsets.all(64),
+
+      children: [for (var value in puzzle.list) 
+        ElevatedButton(
+          onPressed: null,
+          style: ElevatedButton.styleFrom(
+            shape: LinearBorder(),
+          ),
+          child:Text('$value')
+        ),
+        ElevatedButton(
+          onPressed: () {
+            puzzle.rounds++;
+            timer.start();
+          },
+          style: ElevatedButton.styleFrom(
+            shape: LinearBorder(),
+          ),
+          child:Text('')
+        )
+      ]
+    );
+  }
+}
+
+class Puzzle extends StatelessWidget {
+  const Puzzle({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final timer = context.watch<TimerController>();
+    // final puzzle = context.watch<PuzzleController>();
+
+    return Padding(
+      padding: EdgeInsetsGeometry.symmetric(horizontal: 64, vertical: 16),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          TimerSettings(timerController: timer),
+          Settings(),
+          Expanded(child: Canvas())
+        ],
+      ),
+    );
+  }
+}
 
 class PuzzleController extends ChangeNotifier {
   int level = 2; // Nível do Puzzle
@@ -53,66 +108,6 @@ class PuzzleController extends ChangeNotifier {
       shuffleList();
     }
     notifyListeners();
-  }
-}
-
-class Puzzle extends StatelessWidget {
-  const Puzzle({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final timer = context.watch<TimerController>();
-    final puzzle = context.watch<PuzzleController>();
-
-
-    return Padding(
-      padding: EdgeInsetsGeometry.symmetric(horizontal: 64, vertical: 16),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          TimerControls(timerController: timer),
-          Row( /* Game Settings */
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Expanded(child: Text('${puzzle.rounds} jogadas', textAlign: TextAlign.end,)),
-              SizedBox(
-                width: 128,
-                child: TextButton(
-                  onPressed: () async => puzzle.changeLevel(context),  
-                  child: Text('Nível: ${puzzle.level}')
-                ) 
-              ),
-            ]
-          ),
-          Expanded(child: 
-            GridView.count(
-              crossAxisCount: puzzle.level,
-              padding: EdgeInsets.all(64),
-        
-              children: [for (var value in puzzle.list) 
-                ElevatedButton(
-                  onPressed: null,
-                  style: ElevatedButton.styleFrom(
-                    shape: LinearBorder(),
-                  ),
-                  child:Text('$value')
-                ),
-                ElevatedButton(
-                  onPressed: () {
-                    puzzle.rounds++;
-                    timer.start();
-                  },
-                  style: ElevatedButton.styleFrom(
-                    shape: LinearBorder(),
-                  ),
-                  child:Text('')
-                )
-              ]
-            ),
-          )
-        ],
-      ),
-    );
   }
 }
