+++ lib/Puzzle/timer.dart
@@ -1,9 +1,10 @@
-import 'package:flutter/material.dart';
-import 'timer.dart';
+import 'dart:async';
 
-class TimerControls extends StatelessWidget {
+import 'package:flutter/material.dart';
+
+class TimerSettings extends StatelessWidget {
   final TimerController timerController;
-  const TimerControls({
+  const TimerSettings({
     super.key, 
     required this.timerController
   });
@@ -46,5 +47,57 @@ class TimerControls extends StatelessWidget {
         
       ],
     );
+  }
+}
+
+class TimerController extends ChangeNotifier {
+  final refreshRate = 100;
+  final Stopwatch _stopwatch = Stopwatch();
+  Timer? _timer;
+  Duration get elapsed => _stopwatch.elapsed;
+  bool get isRunning => _stopwatch.isRunning;
+
+  void start() {
+    if(_timer != null) return;
+
+    _stopwatch.start();
+    _timer = Timer.periodic(Duration(milliseconds: refreshRate), (_) {
+      notifyListeners();
+    });
+    
+    notifyListeners();
+  }
+
+  void stop() {
+    _stopwatch.stop();
+    _timer?.cancel();
+    _timer = null;
+
+    notifyListeners();
+  }
+
+  void resume() {
+    if (_timer == null) {
+      _timer = Timer.periodic(Duration(milliseconds: refreshRate), (_) {
+        notifyListeners();
+      });
+      _stopwatch.start();
+      notifyListeners();
+    } else {
+      return;
+    }
+  }
+
+  void reset() {
+    _stopwatch.reset();
+    _timer?.cancel();
+    _timer = null;
+    notifyListeners();
+  }
+
+  @override
+  void dispose() {
+    _timer?.cancel();
+    super.dispose();
   }
 }
