+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'app.dart';
-import 'timer.dart';
-import 'puzzle.dart';
+import 'Puzzle/timer.dart';
+import 'Puzzle/puzzle.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
